requirements.txt
streamlit
nltk>=3.5
textblob
newspaper3k
beautifulsoup4
requests
lxml

